requirements.txt
aldryn-apphooks-config==0.6.0
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0
django-appdata==0.3.2
django-classy-tags==2.0.0
django-cms==3.7.4
django-cors-headers==3.5.0
django-filer==1.7.1
django-formtools==2.2
django-js-asset==1.2.2
django-meta==1.7.0
django-mptt==0.11.0
django-parler==2.1
django-polymorphic==2.1.2
django-restful-admin==1.1.3
django-sekizai==2.0.0
django-sortedm2m==3.0.2
django-taggit==1.2.0
django-taggit-autosuggest==0.3.8
django-taggit-templatetags==0.2.5
django-templatetag-sugar==1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-apphook-setup==0.4.1
djangocms-attributes-field==2.0.0
djangocms-blog==1.1.1
djangocms-bootstrap4==1.6.0
djangocms-file==2.4.0
djangocms-googlemap==1.4.0
djangocms-icon==2.0.0
djangocms-link==3.0.0
djangocms-picture==3.0.0
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.10.0
djangocms-video==2.3.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
easy-thumbnails==2.7.1
html5lib==1.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.1
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
ruamel.yaml==0.16.12
six==1.15.0
sqlparse==0.4.1
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.26.2
webencodings==0.5.1